requirements.txt
google-genai>=1.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2 